refactor the merging functionality and use a display item instead
