+++ packages/ui/src/viewer-table/DataTable/SchemaTable/Columns/AllowedValuesColumn/RenderAllowedValues.tsx
@@ -1,0 +1,55 @@
+/*
+ * Copyright (c) 2025 The Ontario Institute for Cancer Research. All rights reserved
+ *
+ *  This program and the accompanying materials are made available under the terms of
+ *  the GNU Affero General Public License v3.0. You should have received a copy of the
+ *  GNU Affero General Public License along with this program.
+ *   If not, see <http://www.gnu.org/licenses/>.
+ *
+ *  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY
+ *  EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES
+ *  OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT
+ *  SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+ *  INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED
+ *  TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS;
+ *  OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER
+ *  IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
+ *  ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+/** @jsxImportSource @emotion/react */
+
+import { css } from '@emotion/react';
+import { SchemaField, SchemaRestrictions } from '@overture-stack/lectern-dictionary';
+
+import { Theme } from '../../../../../theme';
+import { useThemeContext } from '../../../../../theme/ThemeContext';
+import { computeAllowedValuesColumn } from './ComputeAllowedValues';
+
+const linkStyle = (theme: Theme) => css`
+	${theme.typography?.label};
+	color: ${theme.colors.black};
+	cursor: pointer;
+	display: inline-flex;
+	align-items: center;
+	background: none;
+	border: none;
+	padding: 0;
+	margin-top: 4px;
+	text-decoration: underline;
+	&:hover {
+		text-decoration: underline;
+	}
+`;
+
+const pillStyle = (theme: Theme) => ({
+	fontFamily: 'B612 Mono',
+	color: theme.colors.accent_dark,
+	fontWeight: '400',
+	lineHeight: '20px',
+	fontSize: '13px',
+});
+export const renderAllowedValuesColumn = (restrictions: SchemaRestrictions, schemaField: SchemaField) => {
+	const restrictionItems = computeAllowedValuesColumn(restrictions, schemaField);
+	const theme = useThemeContext();
+};
